Meeting_Assistant/requirements.txt
langchain 
langchain_groq 
openai 
pydantic 
pandas 
py-trello
pydub
speechrecognition